#31 backend: delete functionality out of this code, rename file of ActionInstance
