+++ backend/dps_training_k/game/models/__init__.py
@@ -5,4 +5,4 @@ from .personnel import Personnel
 from .user import User
 from .scheduled_event import ScheduledEvent
 from .area import Area
-from .applied_action import *
+from .action_instance import *
